Move desk persistence into provider
